docs(icons): move icon transform logic to util
