+++ apps/docs/src/.vitepress/components/OnyxIconLibrary.vue
@@ -1,64 +1,31 @@
 <script lang="ts" setup>
-import { ICON_CATEGORIES } from "@sit-onyx/icons";
+import { computed, ref } from "vue";
 import OnyxIcon from "~components/OnyxIcon/OnyxIcon.vue";
 import Search from "./Search.vue";
-import { capitalize, computed, ref } from "vue";
-
-const search = ref("");
+import { getEnrichedIconCategoryList } from "../utils-icons";
 
 const ALL_ICONS = import.meta.glob("../../../node_modules/@sit-onyx/icons/src/assets/*.svg", {
   as: "raw",
   eager: true,
 });
+const ENRICHED_ICON_CATEGORY_LIST = getEnrichedIconCategoryList(ALL_ICONS);
 
-const getIconContent = (iconName: string) => {
-  return ALL_ICONS[`../../../node_modules/@sit-onyx/icons/src/assets/${iconName}.svg`];
-};
+const search = ref("");
 
-const getIconContextData = (iconName: string) => {
-  const parts = iconName.split("-");
-  // bell-disabled => Bell Disabled
-  const tooltipName = parts.map((word) => capitalize(word)).join(" ");
-  // bell-disabled => bellDisabled
-  const importName = parts
-    .map((word, index) => {
-      if (index === 0) return word;
-      return capitalize(word);
-    })
-    .join("");
-
-  return { tooltipName, importName, content: getIconContent(iconName) };
-};
-
-// collects the icon contents once for all icons
-const enrichedIconCategories = Object.entries(ICON_CATEGORIES).map(([category, icons]) => ({
-  name: category,
-  icons: icons.map((icon) => ({
-    ...icon,
-    ...getIconContextData(icon.iconName),
-    metadata: {
-      ...icon.metadata,
-      // make the search more reliable
-      aliases: icon.metadata.aliases?.map((alias) => alias.toLowerCase()),
-    },
-  })),
-}));
 const filteredCategories = computed(() => {
   const lowerCaseSearch = search.value.toLowerCase();
-  return enrichedIconCategories
-    .map((category) => {
-      if (category.name.toLowerCase().includes(lowerCaseSearch)) return category;
+  return ENRICHED_ICON_CATEGORY_LIST.map((category) => {
+    if (category.name.toLowerCase().includes(lowerCaseSearch)) return category;
 
-      return {
-        ...category,
-        icons: category.icons.filter(
-          (icon) =>
-            icon.iconName.toLowerCase().includes(lowerCaseSearch) ||
-            icon.metadata.aliases?.some((alias) => alias.includes(lowerCaseSearch)),
-        ),
-      };
-    })
-    .filter((category) => category.icons.length);
+    return {
+      ...category,
+      icons: category.icons.filter(
+        (icon) =>
+          icon.iconName.toLowerCase().includes(lowerCaseSearch) ||
+          icon.metadata.aliases?.some((alias) => alias.includes(lowerCaseSearch)),
+      ),
+    };
+  }).filter((category) => category.icons.length);
 });
 </script>
 
@@ -76,6 +43,7 @@ const filteredCategories = computed(() => {
           :icon="icon.content"
           size="md"
           :title="icon.tooltipName"
+          :color="icon.metadata.deprecated ? 'secondary' : 'currentColor'"
         />
       </div>
     </section>
